+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.2.1] - 2024-07-19
 
 ### Changed
 
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.1.0]: https://github.com/dogmatiq/spruce/releases/tag/v0.1.0
 [0.1.1]: https://github.com/dogmatiq/spruce/releases/tag/v0.1.1
 [0.2.0]: https://github.com/dogmatiq/spruce/releases/tag/v0.2.0
+[0.2.1]: https://github.com/dogmatiq/spruce/releases/tag/v0.2.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD